Fix printf and getsockname arguments for 64-bit
size_t arguments to printf have to use %lu and be cast to unsigned long.
Use socklen_t as the size argument to getsockname, not int.  These fix
warnings with a build on 64-bit RHEL 5.x.

+++ server/config.c
@@ -250,7 +250,7 @@ read_conf_file(void *data, const char *name)
             line = NULL;
             continue;
         } else if (line->count < 4) {
-            warn("%s:%d: parse error", name, lineno);
+            warn("%s:%lu: parse error", name, (unsigned long) lineno);
             goto fail;
         }
